feat: strategy-conditioned routing + tail-risk + risk-weighted sizer
Turns the scored attention map into asset selection, tail-aware scoring, and an
actual position book.

Routing (scoring/routing.py, `trading route`)
- strategy_asset_plan: each strategy -> the assets it has an edge on
- route_symbols_to_strategies: each symbol -> its single best strategy; emits a
  monitor-ready --strategy-map string. Driven by any objective (dot_product, …)

Tail risk (risk/tail.py)
- expected_shortfall (CVaR), cornish_fisher_var (skew/kurtosis-adjusted VaR),
  conditional_drawdown_at_risk, ulcer_index, downside_deviation, tail_ratio,
  omega_ratio, loss_probability
- CVaR baked into MatrixCell + report; new dot_product_cvar objective swaps the
  risk axis from max-drawdown to Expected Shortfall

Risk-weighted sizing (risk/allocation.py)
- risk_weighted_allocation: split one risk budget across candidates by
  equal_risk / risk_parity (1/CVaR) / score (softmax = soft attention) /
  score_cvar (softmax(score)/CVaR); size each to a CVaR-based stop
- the operational form of softmax(Q.K)/risk from the attention framing

+25 tests (275 total). ruff + mypy clean; also drops a redundant int() in sizing.

+++ src/trading_live_claude/risk/sizing.py
@@ -51,7 +51,7 @@ class PositionSizer:
             target = entry - self.target_r * stop_distance
 
         dollar_risk = equity * self.risk_pct
-        shares = max(int(math.floor(dollar_risk / stop_distance)), 0)
+        shares = max(math.floor(dollar_risk / stop_distance), 0)
         return SizingResult(
             shares=shares,
             entry=entry,
